Det: move normalization from Dataset into forward.
By default, channel-wise mean and standard deviation parameters are fit based on one epoch of generated images (instead of using values based on ImageNet). The small ResNet has a default of 0 for mean and 1 for standard deviation; the full ResNet model hasn't been updated yet and won't do normalization properly (a temporary error is thrown if you try to use it).

Having the normalization in the forward method of the model instead of as a transform on the dataset makes it easier to have uniform behavior between training on synthetic images and doing detection on real environment images.