Env: move rendering to base env class.
All other classes now provide the base class with a dictionary that specifies the images to paste and how to get the x, y coordinates (and possibly rotation angle) for each of them.

This removes a lot of redundant code and should make the graphical part of new environments easier.